requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asn1crypto==1.5.1
attrs==25.3.0
blinker==1.9.0
boto3==1.39.0
botocore==1.39.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cohere==5.15.0
colorama==0.4.6
cryptography==45.0.4
distro==1.9.0
fastavro==1.11.1
filelock==3.18.0
Flask==3.1.1
frozenlist==1.7.0
fsspec==2025.5.1
gTTS==2.5.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
lxml==6.0.0
MarkupSafe==3.0.2
multidict==6.6.3
numpy==2.3.1
openai==1.93.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
platformdirs==4.3.8
propcache==0.3.2
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
PyJWT==2.10.1
pyOpenSSL==25.1.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
snowflake==1.6.0
snowflake-connector-python==3.15.0
snowflake._legacy==1.0.1
snowflake.core==1.6.0
sortedcontainers==2.4.0
tokenizers==0.21.2
tomlkit==0.13.3
tqdm==4.67.1
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
webrtcvad-wheels==2.0.14
Werkzeug==3.1.3
yarl==1.20.1
